Add auth content for get-medical-records
