+++ src/applications/static-pages/health-care-manage-benefits/secure-messaging/components/App/index.js
@@ -1,10 +1,14 @@
 // Node modules.
 import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
 // Relative imports.
 import './styles.scss';
 import CallToActionWidget from 'platform/site-wide/cta-widget';
+import MoreInfoAboutBenefits from '../../../components/MoreInfoAboutBenefits';
+import { selectIsCernerPatient } from 'platform/user/selectors';
 
-export const App = () => (
+export const App = ({ isCernerPatient }) => (
   <>
     <CallToActionWidget appId="messaging" setFocus={false} />
     <div>
@@ -306,4 +310,16 @@ export const App = () => (
   </>
 );
 
-export default App;
+App.propTypes = {
+  // From mapStateToProps.
+  isCernerPatient: PropTypes.bool,
+};
+
+const mapStateToProps = state => ({
+  isCernerPatient: selectIsCernerPatient(state),
+});
+
+export default connect(
+  mapStateToProps,
+  null,
+)(App);
